Updated the share icon
Designers in the native space have expressed a concern around .icon-share in the GUI not resembling the native share icon for their particular operating system (iOS and Android).

From a Web perspective the UI icons for our digital brands should not directly imitate any operating system.

I met with Darren and Rob and we have agreed on a more generic share icon which is still a familiar symbol for either audience. This asset will be updated in the GUI soon.