Added icon mask and swatch color placeholder
We’ve added a mask and a swatch color placeholder to each icon in the icon library sketch file. This enables designers to change the colour of an icon using a symbol over-ride. To see this in action open a GUI sketch file, select an icon and look at the symbol over-rides panel. You will notice a Color over-ride which will allow the icon to adopt swatch colors for that particular brand. Thanks to Brad for coming up with this solution.