Updates to color palette
To align with the GUI we’ve removed Comp A and Comp B from the Primary UI palette and we’ve also added a heading color to each brand. Now each brand has a Primary UI Palette and a Seconday Palette. The secondary palette is a bespoke collection of colours specific to the brand. The colours in the Secondary palette are for brand embelishment. These colours do not map to the other brands in the same way that the Primary colours map.